Post-project migration rename. Converting all ember-sinon-sandbox references
