+++ addon-test-support/index.js
@@ -1,1 +1,1 @@
-export { default as setupSinonSandbox } from './setup-sinon-sandbox';
+export { default as setupSinonSinoff } from './setup-sinon-sinoff';
