+++ addon-test-support/setup-global-sinon-sinoff.js
@@ -1,4 +1,4 @@
-import { createSandbox, restoreSandbox } from './sinon-sandbox';
+import { createSandbox, restoreSandbox } from './sinon-sinoff';
 
 /**
  * Allows for creating and restoring a global sinon sandbox per test. This is
@@ -8,7 +8,7 @@ import { createSandbox, restoreSandbox } from './sinon-sandbox';
  * @param {Object} An object containing optional options
  * @public
  */
-export default function setupSinonSandbox(testEnvironment = self.QUnit) {
+export default function setupSinonSinoff(testEnvironment = self.QUnit) {
   testEnvironment.testStart(createSandbox);
   testEnvironment.testDone(restoreSandbox);
 }
