+++ tests/unit/setup-global-sinon-sinoff-test.js
@@ -1,11 +1,11 @@
 import { module, test } from 'qunit';
-import setupSinonSandbox from 'ember-sinon-sandbox/test-support/setup-global-sinon-sandbox';
+import setupSinonSinoff from 'ember-sinon-sinoff/test-support/setup-global-sinon-sinoff';
 import {
   createSandbox,
   restoreSandbox,
-} from 'ember-sinon-sandbox/test-support/sinon-sandbox';
+} from 'ember-sinon-sinoff/test-support/sinon-sinoff';
 
-module('Unit | ember-sinon-sandbox | Setup in testStart/testDone', function() {
+module('Unit | ember-sinon-sinoff | Setup in testStart/testDone', function() {
   test(`configuring setup/restore`, function(assert) {
     assert.expect(4);
 
@@ -24,7 +24,7 @@ module('Unit | ember-sinon-sandbox | Setup in testStart/testDone', function() {
       },
     };
 
-    setupSinonSandbox(qunit);
+    setupSinonSinoff(qunit);
 
     assert.ok(testStartCalled, 'testEnvironment.testStart is called');
     assert.ok(testDoneCalled, 'testEnvironment.testDone is called');
