+++ tests/unit/setup-sinon-sinoff-ember-qunit-test.js
@@ -1,10 +1,12 @@
 import { module, test } from 'qunit';
 import { setupTest } from 'ember-qunit';
-import setupSinonSandbox from 'ember-sinon-sandbox/test-support/setup-sinon-sandbox';
+import setupSinonSinoff from 'ember-sinon-sinoff/test-support/setup-sinon-sinoff';
 
-module('Unit | ember-sinon-sandbox | Validates helper with hooks', function(hooks) {
+module('Unit | ember-sinon-sinoff | Validates helper with hooks', function(
+  hooks
+) {
   setupTest(hooks);
-  setupSinonSandbox(hooks);
+  setupSinonSinoff(hooks);
 
   hooks.beforeEach(function(assert) {
     assert.ok(this.sandbox);
@@ -12,5 +14,5 @@ module('Unit | ember-sinon-sandbox | Validates helper with hooks', function(hook
 
   test('Fake test name', function(assert) {
     assert.expect(1);
-  })
+  });
 });
